core: add block test from dev branch
And drop it from _pkg.dev. Adapted for master.
